requirements.txt
asgiref==3.5.1
Django==3.2.4
django-cors-headers==3.7.0
django-currentuser==0.5.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
Pillow==9.1.0
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
